backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.37
psycopg[binary]==3.2.3
pydantic==2.9.2
segno==1.6.1
python-dotenv==1.0.1
fpdf2==2.8.1
